Move netstandard2.0 compatibility code to its own namespace
