+++ src/QuestPDF.Markdown/Compatibility/CallerArgumentExpressionAttribute.cs
@@ -1,0 +1,14 @@
+#if NETSTANDARD2_0
+namespace System.Runtime.CompilerServices;
+
+[AttributeUsage(AttributeTargets.Parameter)]
+public sealed class CallerArgumentExpressionAttribute : Attribute
+{
+    public CallerArgumentExpressionAttribute(string parameterName)
+    {
+        ParameterName = parameterName;
+    }
+
+    public string ParameterName { get; }
+}
+#endif
