+++ src/QuestPDF.Markdown/Compatibility/IsExternalInit.cs
@@ -1,0 +1,6 @@
+#if NETSTANDARD2_0
+namespace System.Runtime.CompilerServices
+{
+    internal static class IsExternalInit { }
+}
+#endif
