+++ src/QuestPDF.Markdown/Compatibility/MaybeNullWhenAttribute.cs
@@ -1,0 +1,12 @@
+#if NETSTANDARD2_0
+namespace System.Diagnostics.CodeAnalysis
+{
+    [AttributeUsage(AttributeTargets.Parameter)]
+    public sealed class MaybeNullWhenAttribute : Attribute
+    {
+        public MaybeNullWhenAttribute(bool returnValue) => ReturnValue = returnValue;
+        
+        public bool ReturnValue { get; }
+    }
+}
+#endif
